Switched Template-System for Ember to Handlebars and Emblemjs
Yep, that took quite a while. I haven’t ported over the Readme view yet
because neither Handlebars or Emblemjs can render Markdown, but it
won’t be needed anyways in the future. The target will be that the
server only has to server static content (besides the API of course) so
it will be able to chill around much more.